% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubethemes-package.R
\docType{package}
\name{cubethemes-package}
\alias{cubethemes}
\alias{cubethemes-package}
\title{cubethemes: Themes for cubies to create great data visualisations in ggplot2.}
\description{
cubethemes is a package containing themes to be used by staff at Cube Group. These themes allow cubies to create great data visualisations in ggplot2 in accordance with the Cube Group branding guidelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zerogetsamgow/cubethemes}
  \item Report bugs at \url{https://github.com/zerogetsamgow/cubiethemes/repo/issues}
}

}
\author{
\strong{Maintainer}: Samuel Gow \email{samuel.gow@cubegroup.com.au}

}
\keyword{internal}
