% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cube_fonts.R
\name{install_cube_fonts}
\alias{install_cube_fonts}
\title{Install Cube Fonts}
\usage{
install_cube_fonts(font.path = "C:/windows/fonts")
}
\description{
Themes in this package use the Agenda font.
}
\details{
This function allows these to be installed within the package
}
\seealso{
Other cubethemes: 
\code{\link{theme_cube_foundation}()},
\code{\link{theme_cube_green}()},
\code{\link{theme_cube_map}()},
\code{\link{theme_cube_pink}()},
\code{\link{theme_cube_white}()}
}
\concept{cubethemes}
