% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_foundation.R
\name{theme_cube_foundation}
\alias{theme_cube_foundation}
\title{Cube Foundation Theme}
\usage{
theme_cube_foundation(base_size = 15, base_family = "Agenda")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This theme is designed to be a foundation from which to build new
themes, and not meant to be used directly. \code{theme_cube_foundation()}
is a complete theme with only minimal number of elements defined.
It is easier to create new Cube themes by extending this one rather
than \code{\link[ggplot2]{theme_gray}()} or \code{\link[ggplot2]{theme_bw}()},
because those themes define elements deep in the hierarchy.
}
\details{
This theme takes \code{\link[ggthemes]{theme_foundations}()} and sets
base text family and size to \code{Agenda} and \code{15}.
}
\seealso{
Other cubethemes: 
\code{\link{install_cube_fonts}()},
\code{\link{theme_cube_gray}()},
\code{\link{theme_cube_green}()},
\code{\link{theme_cube_grey}()},
\code{\link{theme_cube_map}()},
\code{\link{theme_cube_pink}()},
\code{\link{theme_cube_white}()}
}
\concept{cubethemes}
