% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_grey.R
\name{theme_cube_gray}
\alias{theme_cube_gray}
\title{Cube gray Theme}
\usage{
theme_cube_gray(
  base.colour = cube.grey,
  line.colour = cube.green,
  text.colour = cube.green
)
}
\description{
This theme takes \code{\link[cubethemes]{theme_cube_foundations}()} and sets
background colour to gray and line and text colours to green.
}
\details{
Colours are sourced from \code{cubepalette} and are Cube Brand Guidelines compliant.
}
\seealso{
Other cubethemes: 
\code{\link{install_cube_fonts}()},
\code{\link{theme_cube_foundation}()},
\code{\link{theme_cube_green}()},
\code{\link{theme_cube_grey}()},
\code{\link{theme_cube_map}()},
\code{\link{theme_cube_pink}()},
\code{\link{theme_cube_white}()}
}
\concept{cubethemes}
