% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_orange.R
\name{theme_cube_orange}
\alias{theme_cube_orange}
\title{Cube orange Theme}
\usage{
theme_cube_orange(
  base.colour = cube.lightorange,
  line.colour = cube.darkgreen,
  text.colour = cube.darkgreen
)
}
\description{
This theme takes \code{\link[cubethemes]{theme_cube_foundations}()} and sets
background colour to lightorange and line and text colours to green.
}
\details{
Colours are sourced from \code{cubepalette} and are Cube Brand Guidelines compliant.
}
\seealso{
Other cubethemes: 
\code{\link{install_cube_fonts}()},
\code{\link{theme_cube_foundation}()},
\code{\link{theme_cube_gray}()},
\code{\link{theme_cube_green}()},
\code{\link{theme_cube_grey}()},
\code{\link{theme_cube_map}()},
\code{\link{theme_cube_pink}()},
\code{\link{theme_cube_white}()}
}
\concept{cubethemes}
