% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cube_green.R
\name{theme_cube_green}
\alias{theme_cube_green}
\title{Cube Green Theme}
\usage{
theme_cube_green(
  base.colour = cube.green,
  line.colour = cube.pink,
  text.colour = cube.pink
)
}
\description{
This theme takes \code{\link[cubethemes]{theme_cube_foundations}()} and sets
background colour to green and line and text colours to pink.
}
\details{
Colours are sourced from \code{cubepalette} and are Cube Brand Guidelines compliant.
}
\seealso{
Other cubethemes: 
\code{\link{install_cube_fonts}()},
\code{\link{theme_cube_foundation}()},
\code{\link{theme_cube_map}()},
\code{\link{theme_cube_pink}()},
\code{\link{theme_cube_white}()}
}
\concept{cubethemes}
